refactor(automation): extract pr_review/address_review cores off the omit list
Epic #1809 final omit-reduction wave (13 -> 12).

Split the PR-review and address-review modules into a covered core layer plus
a thin orchestration/CLI layer, mirroring the planner/implementer/ci_driver
pipeline-wrapper conversions of #1820-#1822:

- New hephaestus/automation/pr_review_core.py holds the pure/parse/context
  review cores (run_pr_review_analysis, gather_impl_review_context,
  review_pr_inline). Unit-covered at 87.21%; not on the omit list.
- New hephaestus/automation/address_review_core.py holds the parse/fix/resolve
  cores (_parse_addressed_block, _log_address_parse_error,
  run_address_fix_session, resolve_addressed_threads, _ADDRESS_PARSE_DEFAULT).
  Unit-covered at 80.21%; not on the omit list.
- pr_reviewer.py becomes a thin hephaestus-review-prs pipeline wrapper
  (PR_REVIEW scope) with PRReviewer reduced to an importable placeholder; it is
  REMOVED from [tool.coverage.run].omit and now reports 87.04% coverage.
- address_review.py keeps its live AddressReviewer class + main() and STAYS
  omitted (no console script; live worktree/agent orchestration is
  integration-only). Its parse/fix cores exit omit via address_review_core.

Both wrappers re-export their cores (name as name) so pinned patch sites and
from-import call sites (_review_phase, review_validator, review_thread_resolver)
keep resolving.

Guardrails updated for the 13 -> 12 omit reduction: pyproject.toml omit list,
test_omit_allowlist.py (count 16 -> 12 comment), and
test_orchestration_smoke.py docstring. Patch-seam sweep retargeted the moved
symbols (run_agent_session/gh_pr_resolve_thread/get_repo_slug/
invoke_claude_with_session and the dontAsk/allowed-tools/phase-wiring guards)
to the core modules; AGENTS.md dontAsk rows updated to match. Deleted the
now-obsolete PRReviewer class-method tests (behavior owned by the pipeline
pr_review stage, 72 tests) and rewrote test_pr_reviewer_main.py to the
pipeline-wrapper style.

Also reformatted tests/unit/automation/pipeline/test_coordinator_edges.py to
clear a leftover ruff-format drift from #1946.

Epic omit trajectory: 17 -> 12 (never positive).

Closes #1823
