+++ pyproject.toml
@@ -298,8 +298,6 @@ omit = [
     "hephaestus/automation/loop_repo_manager.py",
     # curses_ui.py uses a live terminal; cannot be unit-tested without a real TTY
     "hephaestus/automation/curses_ui.py",
-    # pr_reviewer.py drives a full Claude session against a live PR; integration-only
-    "hephaestus/automation/pr_reviewer.py",
     # audit_reviewer.py drives a live agent coordinator session; integration-only
     "hephaestus/automation/audit_reviewer.py",
 ]
