+++ tests/unit/automation/test_claude_dontask_policy_docs.py
@@ -12,8 +12,8 @@ EXPECTED_DONTASK_CALLS = {
     "audit_reviewer.py:run_audit_coordinator": "Read,Glob,Grep",
     "review_validator.py:_run_validation_session": "Read,Glob,Grep",
     "comment_difficulty.py:_run_classifier_session": "Read,Glob,Grep",
-    "pr_reviewer.py:run_pr_review_analysis": "Read,Glob,Grep",
-    "address_review.py:run_address_fix_session": "Read,Write,Edit,Glob,Grep,Bash,Task,Skill",
+    "pr_review_core.py:run_pr_review_analysis": "Read,Glob,Grep",
+    "address_review_core.py:run_address_fix_session": "Read,Write,Edit,Glob,Grep,Bash,Task,Skill",
     "_implement_phase.py:ImplementPhase._run_claude_impl_session": (
         "Read,Write,Edit,Glob,Grep,Bash"
     ),
