+++ tests/unit/automation/test_provider_neutral_direct_dispatch.py
@@ -17,10 +17,12 @@ PROVIDER_NEUTRAL_FILES = [
     "hephaestus/automation/implementer.py",
     "hephaestus/automation/_review_phase.py",
     "hephaestus/automation/pr_reviewer.py",
+    "hephaestus/automation/pr_review_core.py",
     "hephaestus/automation/audit_reviewer.py",
     "hephaestus/automation/review_validator.py",
     "hephaestus/automation/comment_difficulty.py",
     "hephaestus/automation/address_review.py",
+    "hephaestus/automation/address_review_core.py",
     "hephaestus/automation/ci_driver.py",
     "hephaestus/automation/ci_fix_orchestrator.py",
     "hephaestus/automation/post_merge_processor.py",
